quiet some data conversions warnings raised by msvc

+++ vm/factor.cpp
@@ -129,7 +129,7 @@ void factor_vm::init_factor(vm_parameters *p)
 	init_callbacks(p->callback_size);
 	load_image(p);
 	init_c_io();
-	init_inline_caching(p->max_pic_size);
+	init_inline_caching((int)p->max_pic_size);
 	if(p->signals)
 		init_signals();
